requirements.txt
python-can
scapy
asyncio
ruamel.yaml
pre-commit
pydantic
python-can
